Tabs

+++ HonoursPython/HoughCircle.py
@@ -11,8 +11,8 @@ class HoughCircle:
         #Constructor method takes in the image.
         def __init__(self, img):
             #Get the image.
-            self.img = img
-            #self.img = np.uint8(img)
+            #self.img = img
+            self.img = np.uint8(img)
             
             #Removes noise.
             self.img = cv2.medianBlur(self.img, 5)
